Avoid duplicating db data during saves
In rare circumstances (perhaps when Redis keys have been flushed), the
server logic will start a new Wikipedia paragraph fetch, which leads
to database saves, which will result in duplicate data. There is now
a check to see if the Wikipedia title exists. If so, the server will
avoid the database save.