Change host for lobby image
- To leverage higher simultaneous connections for multiple hosts, change
the lobby image from being hosted on S3 to Imgur. When the lobby is
shown, there are many images needing to be loaded. The loading of the
small thumbnails for the big scenes will easily fill up the block of
6 simultaneous connections to S3, so let's try to have the browser also
simultaneously load the lobby image from Imgur and try to avoid the
broken "warpy" texture mapping when images aren't ready for a3-threejs
to display all asssets within 10 seconds.