Change position and size of 'info' icon
- To keep things looking really professional, we should strive for
consistency in size and look of the iconography. Our brains love
repetition and variation, so keeping all the UI icons with the same grid
spacing and circular background yet having different icon symbols will
please our brains which tries to organize visual elements in
a gestalt-like attempt to create order out of chaos.